+++ src/RcppExports.cpp
@@ -48,8 +48,8 @@ BEGIN_RCPP
 END_RCPP
 }
 // loglik_res
-Rcpp::List loglik_res(const Eigen::SparseMatrix<double> A, const double ldetB, const double psi_r, const Eigen::SparseMatrix<double> H, Eigen::VectorXd Y, const Eigen::Map<Eigen::MatrixXd> X, const Eigen::MappedSparseMatrix<double> Z, const Eigen::Map<Eigen::MatrixXd> XtX, const Eigen::Map<Eigen::MatrixXd> XtZ, const Eigen::MappedSparseMatrix<double> ZtZ, const Eigen::Map<Eigen::VectorXd> XtY, const Eigen::Map<Eigen::VectorXd> ZtY, const bool get_val, const bool get_score, const bool get_inf);
-RcppExport SEXP _reconf_loglik_res(SEXP ASEXP, SEXP ldetBSEXP, SEXP psi_rSEXP, SEXP HSEXP, SEXP YSEXP, SEXP XSEXP, SEXP ZSEXP, SEXP XtXSEXP, SEXP XtZSEXP, SEXP ZtZSEXP, SEXP XtYSEXP, SEXP ZtYSEXP, SEXP get_valSEXP, SEXP get_scoreSEXP, SEXP get_infSEXP) {
+Rcpp::List loglik_res(const Eigen::SparseMatrix<double> A, const double ldetB, const double psi_r, const Eigen::SparseMatrix<double> H, Eigen::VectorXd Y, const Eigen::Map<Eigen::MatrixXd> X, const Eigen::MappedSparseMatrix<double> Z, const Eigen::Map<Eigen::MatrixXd> XtX, const Eigen::Map<Eigen::MatrixXd> XtZ, const Eigen::MappedSparseMatrix<double> ZtZ, const Eigen::Map<Eigen::VectorXd> XtY, const Eigen::Map<Eigen::VectorXd> ZtY, const bool get_val, const bool get_score, const bool get_inf, const bool expected);
+RcppExport SEXP _reconf_loglik_res(SEXP ASEXP, SEXP ldetBSEXP, SEXP psi_rSEXP, SEXP HSEXP, SEXP YSEXP, SEXP XSEXP, SEXP ZSEXP, SEXP XtXSEXP, SEXP XtZSEXP, SEXP ZtZSEXP, SEXP XtYSEXP, SEXP ZtYSEXP, SEXP get_valSEXP, SEXP get_scoreSEXP, SEXP get_infSEXP, SEXP expectedSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
@@ -68,7 +68,8 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< const bool >::type get_val(get_valSEXP);
     Rcpp::traits::input_parameter< const bool >::type get_score(get_scoreSEXP);
     Rcpp::traits::input_parameter< const bool >::type get_inf(get_infSEXP);
-    rcpp_result_gen = Rcpp::wrap(loglik_res(A, ldetB, psi_r, H, Y, X, Z, XtX, XtZ, ZtZ, XtY, ZtY, get_val, get_score, get_inf));
+    Rcpp::traits::input_parameter< const bool >::type expected(expectedSEXP);
+    rcpp_result_gen = Rcpp::wrap(loglik_res(A, ldetB, psi_r, H, Y, X, Z, XtX, XtZ, ZtZ, XtY, ZtY, get_val, get_score, get_inf, expected));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -91,8 +92,8 @@ BEGIN_RCPP
 END_RCPP
 }
 // loglik_res_n
-Rcpp::List loglik_res_n(const Eigen::Map<Eigen::MatrixXd> K, const Eigen::Map<Eigen::VectorXd> psi, const Eigen::Map<Eigen::VectorXd> Y, const Eigen::Map<Eigen::MatrixXd> X, const bool get_val, const bool get_score, const bool get_inf);
-RcppExport SEXP _reconf_loglik_res_n(SEXP KSEXP, SEXP psiSEXP, SEXP YSEXP, SEXP XSEXP, SEXP get_valSEXP, SEXP get_scoreSEXP, SEXP get_infSEXP) {
+Rcpp::List loglik_res_n(const Eigen::Map<Eigen::MatrixXd> K, const Eigen::Map<Eigen::VectorXd> psi, const Eigen::Map<Eigen::VectorXd> Y, const Eigen::Map<Eigen::MatrixXd> X, const bool get_val, const bool get_score, const bool get_inf, const bool expected);
+RcppExport SEXP _reconf_loglik_res_n(SEXP KSEXP, SEXP psiSEXP, SEXP YSEXP, SEXP XSEXP, SEXP get_valSEXP, SEXP get_scoreSEXP, SEXP get_infSEXP, SEXP expectedSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
@@ -103,7 +104,8 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< const bool >::type get_val(get_valSEXP);
     Rcpp::traits::input_parameter< const bool >::type get_score(get_scoreSEXP);
     Rcpp::traits::input_parameter< const bool >::type get_inf(get_infSEXP);
-    rcpp_result_gen = Rcpp::wrap(loglik_res_n(K, psi, Y, X, get_val, get_score, get_inf));
+    Rcpp::traits::input_parameter< const bool >::type expected(expectedSEXP);
+    rcpp_result_gen = Rcpp::wrap(loglik_res_n(K, psi, Y, X, get_val, get_score, get_inf, expected));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -111,9 +113,9 @@ END_RCPP
 static const R_CallMethodDef CallEntries[] = {
     {"_reconf_Psi_from_H_cpp", (DL_FUNC) &_reconf_Psi_from_H_cpp, 2},
     {"_reconf_loglik", (DL_FUNC) &_reconf_loglik, 14},
-    {"_reconf_loglik_res", (DL_FUNC) &_reconf_loglik_res, 15},
+    {"_reconf_loglik_res", (DL_FUNC) &_reconf_loglik_res, 16},
     {"_reconf_loglik_n", (DL_FUNC) &_reconf_loglik_n, 8},
-    {"_reconf_loglik_res_n", (DL_FUNC) &_reconf_loglik_res_n, 7},
+    {"_reconf_loglik_res_n", (DL_FUNC) &_reconf_loglik_res_n, 8},
     {NULL, NULL, 0}
 };